Re-use directory for all cargo test sub-processes
This puts a bunch of infrastructure in place which lets us re-use the
same numbered directory for all `cargo test` sub-processes.  This is
useful since there are multiple binaries which are run, e.g. one for
all unittests, one for each integration tests, one for each doctest.
Not doing this would quickly cycle through numbered directories and
make it very hard to find the most recent run without limiting the
tests to just one binary.